# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auction Platform API",
        "description": "Public REST API for auction platform integration.",
        "contact": {
            "email": "api@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auctions": {
            "get": {
                "tags": [
                    "Auctions"
                ],
                "summary": "List active auctions",
                "operationId": "999974a75bacd75345d926a7ec7d84ba",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ending_soon",
                                "newest",
                                "price_asc",
                                "price_desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated auction list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuctionCollection"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auctions/{auction}": {
            "get": {
                "tags": [
                    "Auctions"
                ],
                "summary": "Get auction details",
                "operationId": "58cf098816e90e19984331d41e4f36a6",
                "parameters": [
                    {
                        "name": "auction",
                        "in": "path",
                        "description": "Auction ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auction details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuctionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/auth/token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create new API token",
                "operationId": "91fabd5c5e0f3fff64ee77e8110689c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string"
                                    },
                                    "abilities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "abilities": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/v1/auth/revoke": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke current API token",
                "operationId": "71e34edaff82485f03607c2630f5e141",
                "responses": {
                    "200": {
                        "description": "Token revoked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auctions/{auction}/bids": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "List bids for an auction",
                "operationId": "69bfa59678f3cf915576665260185030",
                "parameters": [
                    {
                        "name": "auction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bids"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bids"
                ],
                "summary": "Place a bid",
                "operationId": "7c9c4c00f12a057327cc2dd917cd6328",
                "parameters": [
                    {
                        "name": "auction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bid placed successfully"
                    },
                    "400": {
                        "description": "Validation Error"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuctionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "current_price": {
                        "type": "number"
                    },
                    "starting_price": {
                        "type": "number"
                    },
                    "reserve_met": {
                        "type": "boolean"
                    },
                    "reserve_price": {
                        "type": "number",
                        "nullable": true
                    },
                    "min_bid_increment": {
                        "type": "number"
                    },
                    "next_minimum_bid": {
                        "type": "number"
                    },
                    "bid_count": {
                        "type": "integer"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "condition": {
                        "type": "string"
                    },
                    "condition_label": {
                        "type": "string"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "time_remaining": {
                        "type": "string"
                    },
                    "seller": {
                        "type": "object"
                    },
                    "brand": {
                        "type": "object",
                        "nullable": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "links": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AuctionCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuctionResource"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}